Revert "Order overview + More seamless uploading experience"
