+++ SPFAdminSystem/SPFAdminSystem/Database/ProductFiles/ProductService.cs
@@ -6,15 +6,9 @@ using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
 using OfficeOpenXml;
-using System.Data.Entity.Core.Common.CommandTrees.ExpressionBuilder;
 
 namespace SPFAdminSystem.Database.ProductFiles
 {
-    public class ProductScore
-    {
-        public double score { get; set; }
-        public Product _product = new();
-    }
     public class ProductService : IProductService
     {
         private readonly DataContext _context;
@@ -25,10 +19,7 @@ namespace SPFAdminSystem.Database.ProductFiles
         }
 
         public List<Product> Products { get; set; } = new List<Product>();
-        public List<Product> UnknownProducts { get; set; } = new List<Product>();
         public List<Mapping> Mappings { get; set; } = new List<Mapping>();
-        public List<Product> MatchSuggestions { get; set; } = new List<Product>();
-
 
         public async Task LoadProducts()
         {
@@ -43,15 +34,13 @@ namespace SPFAdminSystem.Database.ProductFiles
         public async Task<Product> GetSingleProduct(string productId)
         {
             Product? product = await _context.Products.FindAsync(productId);
-
+            
             if (product == null)
             {
                 throw new Exception("no product here");
             }
             return product;
         }
-
-
 
         public async Task CreateOrUpdateProduct(Product product)
         {
@@ -60,8 +49,7 @@ namespace SPFAdminSystem.Database.ProductFiles
             {
                 //create
                 _context.Products.Add(product);
-            }
-            else
+            } else
             {
                 // update /TOTEST
                 dbProduct.ProductId = product.ProductId;
@@ -69,7 +57,7 @@ namespace SPFAdminSystem.Database.ProductFiles
             await _context.SaveChangesAsync();
         }
 
-
+        
 
         public async Task InsertExcelProducts(string fileName)
         {
@@ -93,7 +81,6 @@ namespace SPFAdminSystem.Database.ProductFiles
                     prod.Ordered = Convert.ToInt32(worksheet.Cells[row, 8].Value);
                     products.Add(prod);
                 }
-                package.Dispose();
             }
             foreach (Product product in products)
             {
@@ -108,23 +95,17 @@ namespace SPFAdminSystem.Database.ProductFiles
             if (dbMapping == null)
             {
                 //create
-
                 _context.Mappings.Add(mapping);
             }
-/*            else if (dbMapping.ProductIdMapping.ToString() == "N/A")
-            {
-
-            }*/
             else
             {
                 // update /TOTEST
                 dbMapping.ProductIdMapping = mapping.ProductIdMapping;
-                
             }
             await _context.SaveChangesAsync();
         }
 
-        public async Task InsertExcelMapping(string fileName)
+            public async Task InsertExcelMapping(string fileName)
         {
             List<Mapping> mappings = new();
             var FilePath = $"{Directory.GetCurrentDirectory()}{@"\wwwroot"}" + "\\" + fileName;
@@ -138,9 +119,9 @@ namespace SPFAdminSystem.Database.ProductFiles
                 for (int row = 2; row <= rowCount; row++)
                 {
                     Mapping Map = new Mapping();
-                    
-                    //create
-                    Map.ProductIdMapping = (worksheet.Cells[row, 2].Value == null || worksheet.Cells[row, 2].Value == string.Empty) ? $"N/A{row}" : worksheet.Cells[row, 2].Value.ToString();
+                   
+                        //create
+                    Map.ProductIdMapping = worksheet.Cells[row, 2].Value == null ? string.Empty : worksheet.Cells[row, 2].Value.ToString();
                     Map.TitleGWS = worksheet.Cells[row, 3].Value.ToString();
                     Map.Barcode = worksheet.Cells[row, 1].Value == null ? string.Empty : worksheet.Cells[row, 1].Value.ToString();
                     Map.Target = Convert.ToInt32(worksheet.Cells[row, 5].Value);
@@ -148,9 +129,7 @@ namespace SPFAdminSystem.Database.ProductFiles
                     Map.PackSize = Convert.ToInt32(worksheet.Cells[row, 6].Value);
                     mappings.Add(Map);
                     Console.WriteLine(Map.ProductIdMapping);
-                    
                 }
-                package.Dispose();
             }
             foreach (Mapping map in mappings)
             {
@@ -178,8 +157,7 @@ namespace SPFAdminSystem.Database.ProductFiles
             {
                 Console.WriteLine("could not find product with id of : " + mapping.ProductIdMapping);
             }
-            else
-            {
+            else { 
                 dbProduct.TitleGWS = mapping.TitleGWS;
                 dbProduct.Barcode = mapping.Barcode;
                 dbProduct.Packsize = mapping.PackSize;
@@ -190,8 +168,7 @@ namespace SPFAdminSystem.Database.ProductFiles
             await _context.SaveChangesAsync();
         }
 
-        public async Task JoinMappingToProducts()
-        {
+        public async Task JoinMappingToProducts (){ 
             await LoadMappings();
             await LoadProducts();
             foreach (Mapping map in Mappings)
@@ -206,175 +183,9 @@ namespace SPFAdminSystem.Database.ProductFiles
                 throw new KeyNotFoundException("product not found");
             return prod;
         }
-
-        public async Task<List<Product>> GetUnknownProducts(string fileName)
-        {
-
-            /*Calls function to assure products from database is in List<Product> Products*/
-            await LoadMappings();
-            await LoadUnknownProducts(fileName);
-            return UnknownProducts;
-        }
-
-
-        public async Task LoadUnknownProducts(string fileName)
-        {
-
-            /*EPPLUS - Excel functionality*/
-            var FilePath = $"{Directory.GetCurrentDirectory()}{@"\wwwroot"}" + "\\" + @fileName;
-            FileInfo fileInfo = new FileInfo(FilePath);
-            ExcelPackage.LicenseContext = LicenseContext.NonCommercial;
-            using (ExcelPackage package = new ExcelPackage(fileInfo))
-            {
-                ExcelWorksheet worksheet = package.Workbook.Worksheets.FirstOrDefault();
-                int colCount = worksheet.Dimension.End.Column;
-                int rowCount = worksheet.Dimension.End.Row;
-                for (int row = 3; row <= rowCount; row++)
-                {
-                    Product prod = new Product();
-                    int isFound = 0;
-                    foreach (Mapping mapping in Mappings)
-                    {
-                        string ExcelBarcode = worksheet.Cells[row, 6].Value.ToString();
-                        if (mapping.Barcode.ToString() == ExcelBarcode)
-                        {
-                            isFound++;
-                            break;
-                        }
-
-                    }
-                    if (isFound == 0)
-                    {
-                        prod.ProductId = worksheet.Cells[row, 5].Value.ToString();
-                        prod.Barcode = worksheet.Cells[row, 6].Value.ToString();
-                        prod.TitleGWS = worksheet.Cells[row, 7].Value.ToString();
-                        prod.Packsize = Convert.ToInt32(worksheet.Cells[row, 9].Value);
-                        prod.OrderPrice = Convert.ToDouble(worksheet.Cells[row, 16].Value);
-                        UnknownProducts.Add(prod);
-                    }
-
-                }
-                Console.WriteLine("Unknown Products added");
-            }
-
-        }
-        public async Task<List<Product>> GetMatchSuggestions(Product product)
-        {
-            List<ProductScore> prodScore = new List<ProductScore>();
-            List<ProductScore> sortedProdScore = new();
-
-            foreach (Mapping map in Mappings)
-            {
-                ProductScore prod = new();
-                prod.score = NameMatch(product.TitleGWS, map.TitleGWS);
-                prod._product.Target = map.Target;
-                prod._product.TitleGWS = map.TitleGWS;
-                prod._product.Barcode = map.Barcode;
-                prod._product.Packsize = map.PackSize;
-                prod._product.MinOrder = map.MinOrder;
-                prodScore.Add(prod);
-            }
-
-            ProductScore newProduct = new();
-            newProduct._product.TitleGWS = "new";
-            sortedProdScore = prodScore.OrderByDescending(x => x.score).ToList();
-
-            MatchSuggestions.Add(sortedProdScore[0]._product);
-            MatchSuggestions.Add(sortedProdScore[1]._product);
-            MatchSuggestions.Add(sortedProdScore[2]._product);
-            MatchSuggestions.Add(sortedProdScore[3]._product);
-            MatchSuggestions.Add(sortedProdScore[4]._product);
-            MatchSuggestions.Add(newProduct._product);
-
-
-            return MatchSuggestions;
-
-        }
-
-
-
-        static double NameMatch(string a, string b)
-        {
-            bool ignoreChecked = false;
-            string[] ignoreWords = { " the ", " of ", " a ", " to ", " & " };
-            /*a = a.ToLower();
-            b = b.ToLower();*/
-
-            bool bigScore = true;
-            double stringLength = 0;
-            string partString = "";
-            string[] searchString = a.Split();
-            string[] mappingString = b.Split();
-            double score = 0;
-            string charString = "";
-            
-            for (int i = 0; i < searchString.Length; i++)
-            {
-                charString = searchString[i];
-                for (int j = 0; j < charString.Length; j++)
-                {
-                    if (charString[j] == ':')
-                    {
-                        charString = charString.Remove(j, 1);
-                    }
-                }
-                searchString[i] = charString;
-            }
-            for (int i = 0; i < mappingString.Length; i++)
-            {
-                charString = mappingString[i];
-                for (int j = 0; j < charString.Length; j++)
-                {
-                    if (charString[j] == ':')
-                    {
-                        charString = charString.Replace(":", "");
-                    }
-                }
-                mappingString[i] = charString;
-            }
-            b = string.Join(" ", mappingString);
-            a = string.Join(" ", searchString);
-
-            stringLength = searchString.Length;
-            stringLength = Math.Floor(stringLength/2);
-            
-            for(int x = 0; x < stringLength; x++)
-            {
-                partString += searchString[x] + " ";
-            }
-
-            if(b.Contains(partString, StringComparison.OrdinalIgnoreCase) && bigScore == true)
-            {
-                score = 1;
-                bigScore = false;
-            }
-            else
-            {
-                bigScore = false;
-            }
-
-            for (int i = 0; i < searchString.Length; i++)
-            {
-                ignoreChecked = false;
-                for (int j = 0; j < ignoreWords.Length; j++)
-                {
-                    if (searchString[i].Contains(ignoreWords[j], StringComparison.OrdinalIgnoreCase))
-                    {
-                        ignoreChecked = true;
-                        break;
-                    }
-                }
-
-                if (b.Contains(searchString[i], StringComparison.OrdinalIgnoreCase) && ignoreChecked == false)
-                {
-                    score++;
-                }
-            }
-
-            return score;
-        }
-
-
-
     }
+
+    
+    
+
 }
