+++ SPFAdminSystem/SPFAdminSystem/Pages/BlazorTemplatePages/Index.razor
@@ -1,6 +1,7 @@
 ﻿@page "/"
 
 <PageTitle>Index</PageTitle>
+
 <h1>Hello, world!</h1>
 
 Welcome to your new app.
