--- SPFAdminSystem/SPFAdminSystem/Pages/BlazorTemplatePages/MatchProducts.razor.css
@@ -1,242 +1,0 @@
-﻿#topSpacer {
-    margin: 0;
-    padding: 0;
-    padding-bottom: 50px;
-}
-
-.container {
-    max-width: 100%;
-    display: flex;
-}
-
-#productList {
-    margin-top: 100px;
-    margin-right: 2%;
-    width: 300px;
-    min-width: 300px;
-}
-
-#productListItems {
-    font-size: 14px;
-    height: 500px;
-    overflow: scroll;
-    padding: 0;
-    border: 2px solid black;
-    -webkit-overflow-scrolling: auto;
-    overflow-x: hidden;
-    cursor: pointer;
-}
-
-li {
-    color: black;
-    height: 50px;
-    font-weight: 500;
-    padding: 10px 20px;
-    border-bottom: 1px solid black;
-}
-
-    li:last-child {
-        border-bottom: none;
-    }
-
-#listTitle {
-    height: 30px;
-    padding: 0px;
-    padding-left: 5px;
-    font-weight: 700;
-    background: #c3c3c3;
-    border-left: 2px solid black;
-    border-right: 2px solid black;
-    border-top: 2px solid black;
-    list-style-type: none;
-    margin-left: 0;
-    margin-right: 0;
-    padding-left: 10px;
-    padding-right: 0;
-    font-size: 20px;
-}
-
-#matchProduct {
-    height: 700px;
-    width: 100%;
-    max-width: 1200px;
-    border: 2px solid black;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-#leftItem {
-    height: 100%;
-    width: 50%;
-    border-right: 1px solid black;
-}
-
-#rightItem {
-    height: 100%;
-    width: 50%;
-    border-left: 1px solid black;
-}
-
-#itemTitle {
-    height: 75px;
-    padding-left: 15px;
-    padding-top: 10px;
-}
-
-#currentInfoTitle {
-    margin-top: 25px;
-}
-
-#leftMatchItemsFlex {
-    display: flex;
-    height: 100%;
-    width: 100%;
-    height: 500px;
-    max-width: 95%;
-    border: 2px solid black;
-    margin: auto;
-    overflow: scroll;
-    -webkit-overflow-scrolling: auto;
-    overflow-x: hidden;
-}
-
-#currentInfo {
-    height: 100%;
-}
-
-#currentInfoValues {
-    text-align: right;
-    height: 100%;
-    width: 150%;
-}
-
-#matchListLeft {
-    height: 100px;
-    width: 100%;   
-}
-
-#itemLi, #itemLiTitles, #itemValueLi {
-    color: black;
-    border-bottom: none;
-}
-
-#itemValueLi {
-    text-align: right;
-}
-
-#browseBtn {
-    margin-left: 1%;
-    padding-top: 40px;
-}
-
-#browseBtnBlue {
-    font-weight: 700;
-    color: white;
-    background: #0288d1;
-    width: 175px;
-    margin-bottom: 15px;
-}
-
-#listButtons {
-    margin: 0;
-    padding: 0;
-    border: 0;
-    width: 105%;
-    height: 105%;
-    text-align: left;
-}
-
-#itemLiTitles {
-    list-style-type: none;
-    margin-left: 0;
-    margin-right: 0;
-    padding-left: 10;
-    padding-right: 0;
-    text-align: left;
-}
-
-#itemLi {
-    text-align: right;
-}
-
-#ignoreApplyButtons {
-    padding-right: 30px;
-    padding-top: 12px;
-    display: flex;
-    justify-content: space-evenly;
-}
-
-#igApButtons {
-    width: 70px;
-}
-
-#applyButton, #acceptButton, #ignoreButton {
-    text-align: center;
-    width: 70px;
-}
-
-#rightAuxItems {
-    display: flex;
-    flex: auto;
-    padding-top: 25px;
-}
-
-#dropdownId {
-    margin-left: auto;
-    margin-right: auto;
-    padding-left: 0px;
-    padding-right: 0px;
-    font-weight: 700;
-}
-
-#matches {
-    width: 475px;
-    height: 30px;
-    padding-left: 10px;
-}
-
-#searchArea {
-    height: 30px;
-    width: 100px;
-}
-
-::placeholder { /* Chrome, Firefox, Opera, Safari 10.1+ */
-    color: black;
-    opacity: 1; /* Firefox */
-}
-
-:-ms-input-placeholder { /* Internet Explorer 10-11 */
-    color: black;
-}
-
-::-ms-input-placeholder { /* Microsoft Edge */
-    color: black;
-}
-
-#inputEditsTarget, #inputEditsInHouseTitle {
-    border: none;
-    text-align: right;
-    padding-top: 10px;
-    padding-bottom: 12px;
-    margin-right: 20px;
-}
-
-#inputEditsTarget {
-    width: 40px;
-}
-
-#inputEditsInHouseTitle {
-    width: 100px;
-}
-
-@media (max-width: 1972px) {
-    .container {
-        display: block;
-    }
-}
-
-#changeOverview{
-    margin-left: 15px;
-    margin-top: 50px;
-}
