+++ SPFAdminSystem/SPFAdminSystem/Pages/DatabasePages/Products.razor
@@ -66,11 +66,13 @@ else
     public string Title { get; private set; }
     private List<Product> products = new();
 
+
     protected override async Task OnInitializedAsync()
     {
         await _productDb.LoadProducts();
         products = _productDb.GetProducts();
     }
+    
 
     public async Task InsertProduct() {
 
@@ -84,5 +86,9 @@ else
         products.Add(p);
         Console.WriteLine("Product Added");
     }
+
+    
+
+
 }
