--- SPFAdminSystem/SPFAdminSystem/Pages/Overview.razor.css
@@ -1,3 +1,0 @@
-#table-div {
-    width: 75%;
-}
