+++ SPFAdminSystem/SPFAdminSystem/Program.cs
@@ -28,8 +28,6 @@ builder.Services.AddServerSideBlazor();
 builder.Services.AddScoped<ProtectedSessionStorage>();
 builder.Services.AddScoped<AuthenticationStateProvider, CustomAuthenticationStateProvider>();
 builder.Services.AddSingleton<WeatherForecastService>();
-builder.Services.AddBlazorTable();
-builder.Services.AddBlazorDownloadFile();
 
 
 var folder = Environment.SpecialFolder.LocalApplicationData;
