+++ SPFAdminSystem/SPFAdminSystem/SPFAdminSystem.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -12,15 +12,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Remove="Files\Pricelist.xlsx" />
-  </ItemGroup>
-
-  <ItemGroup>
     <Content Include="DataAccessLibrary\.vs\VSWorkspaceState.json" />
-    <Content Include="Files\Pricelist.xlsx">
-      <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
   </ItemGroup>
 
   <ItemGroup>
@@ -33,15 +25,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Aspose.Cells" Version="22.11.0" />
     <PackageReference Include="BCrypt.Net-Next.StrongName" Version="4.0.3" />
-    <PackageReference Include="BlazorDownloadFile" Version="2.4.0.2" />
-    <PackageReference Include="BlazorDownloadFileFast" Version="0.2.0" />
-    <PackageReference Include="BlazorTable" Version="1.17.0" />
     <PackageReference Include="Dapper" Version="2.0.123" />
     <PackageReference Include="EPPlus" Version="6.1.1" />
-    <PackageReference Include="FreeSpire.XLS" Version="12.7.0" />
-    <PackageReference Include="IronXL.Excel" Version="2022.11.10251" />
     <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="6.0.9" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="6.0.9" />
     <PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="6.0.9" />
@@ -51,8 +37,7 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Office.Interop.Excel" Version="15.0.4795.1001" />
-    <PackageReference Include="SoloX.BlazorJsBlob" Version="1.0.0" />
+    <PackageReference Include="Radzen.Blazor" Version="4.3.8" />
     <PackageReference Include="System.Data.SqlClient" Version="4.8.5" />
     <PackageReference Include="System.Data.SQLite" Version="1.0.116" />
   </ItemGroup>
