--- SPFAdminSystem/SPFAdminSystem/Shared/Loading.razor
@@ -1,8 +1,0 @@
-﻿<div class="position-fixed bg-black bg-opacity-75" style="z-index: 9999; margin: 0; padding: 0;top: 0; left: 0; width: 100vw; height: 100vh; display: flex; flex-direction: column; justify-content: center; align-items: center;">
-    <div class="spin"></div>
-    <p style="color: white; font-size: 32px; margin-top: 2%;">@Text</p>
-    <p style="color: white;">Please do not refresh</p>
-</div>
-@code {
-    [Parameter] public string Text { get; set; }
-}
