--- SPFAdminSystem/SPFAdminSystem/Shared/Loading.razor.css
@@ -1,34 +1,0 @@
-@keyframes spinner {
-    0%
-
-{
-    transform: translate3d(-50%, -50%, 0) rotate(0deg);
-}
-
-100% {
-    transform: translate3d(-50%, -50%, 0) rotate(360deg);
-}
-
-}
-
-.spin {
-    padding: 0;
-    margin: 0;
-
-}
-
-.spin::before {
-    animation: 1.5s linear infinite spinner;
-    animation-play-state: inherit;
-    border: solid 5px #cfd0d1;
-    border-bottom-color: #1c87c9;
-    border-radius: 50%;
-    content: "";
-    height: 40px;
-    width: 40px;
-    position: fixed;
-    top: 45vh;
-    left: 50vw;
-    transform: translate3d(-50%, -50%, 0);
-    will-change: transform;
-}
