+++ SPFAdminSystem/SPFAdminSystem/Shared/PopupModals/DeleteUserPopup.razor
@@ -31,6 +31,7 @@
                         </tr>
                     </tbody>
                 </table>
+                <p class="small opacity-75 pt-5 pb-0 mb-0">Deleting a user stores the user for 14 days, before being deleted permanently (functionality not implemented yet)</p>
             </div>
             <div class="modal-footer flex-row justify-content-around">
                 <button type="button" class="btn btn-danger" @onclick="() => deleteUser(user)">Delete User</button>
